requirements.txt
appdirs==1.4.3
asgiref==3.2.7
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-import-export==2.4.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
filelock==3.0.12
google-api-core==1.22.2
google-api-python-client==1.11.0
google-auth==1.21.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.1.1
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.13
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
